refactor: move snackbar logic to HeaderContainer component
